CDMODULE/requirements.txt
annotated-types==0.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
h11==0.14.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
pymodbus==3.7.0
PyYAML==6.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
